refactor: move some utilities to ui package
