+++ packages/core/src/controllers/ConnectorController.ts
@@ -1,7 +1,7 @@
 import { subscribeKey as subKey } from 'valtio/vanilla/utils'
 import { proxy, ref, snapshot } from 'valtio/vanilla'
 import type { Connector, EmailConnector } from '../utils/TypeUtil.js'
-import { getCommonThemeVariables } from '@web3modal/common'
+import { getW3mThemeVariables } from '@web3modal/common'
 import { OptionsController } from './OptionsController.js'
 import { ThemeController } from './ThemeController.js'
 
@@ -35,18 +35,18 @@ export const ConnectorController = {
     if (connector.id === 'w3mEmail') {
       const emailConnector = connector as EmailConnector
       const optionsState = snapshot(OptionsController.state) as typeof OptionsController.state
+      const themeMode = ThemeController.getSnapshot().themeMode
+      const themeVariables = ThemeController.getSnapshot().themeVariables
+
       emailConnector?.provider?.syncDappData?.({
         metadata: optionsState.metadata,
         sdkVersion: optionsState.sdkVersion,
         projectId: optionsState.projectId
       })
       emailConnector.provider.syncTheme({
-        themeMode: ThemeController.getSnapshot().themeMode,
-        themeVariables: ThemeController.getSnapshot().themeVariables,
-        w3mThemeVariables: getCommonThemeVariables(
-          ThemeController.getSnapshot().themeVariables,
-          ThemeController.getSnapshot().themeMode
-        )
+        themeMode,
+        themeVariables,
+        w3mThemeVariables: getW3mThemeVariables(themeVariables, themeMode)
       })
     }
   },
