+++ packages/core/src/controllers/ThemeController.ts
@@ -1,7 +1,7 @@
 import { proxy, subscribe as sub, snapshot } from 'valtio/vanilla'
 import type { ThemeMode, ThemeVariables } from '../utils/TypeUtil.js'
 import { ConnectorController } from './ConnectorController.js'
-import { getCommonThemeVariables } from '@web3modal/common'
+import { getW3mThemeVariables } from '@web3modal/common'
 import type { W3mThemeVariables } from '@web3modal/common'
 
 // -- Types --------------------------------------------- //
@@ -28,16 +28,17 @@ export const ThemeController = {
 
   setThemeMode(themeMode: ThemeControllerState['themeMode']) {
     state.themeMode = themeMode
+
     try {
       const emailConnector = ConnectorController.getEmailConnector()
 
       if (emailConnector) {
+        const themeMode = ThemeController.getSnapshot().themeMode
+        const themeVariables = ThemeController.getSnapshot().themeVariables
+
         emailConnector.provider.syncTheme({
-          themeMode: ThemeController.getSnapshot().themeMode,
-          w3mThemeVariables: getCommonThemeVariables(
-            ThemeController.getSnapshot().themeVariables,
-            themeMode
-          )
+          themeMode,
+          w3mThemeVariables: getW3mThemeVariables(themeVariables, themeMode)
         })
       }
     } catch {
@@ -48,15 +49,15 @@ export const ThemeController = {
 
   setThemeVariables(themeVariables: ThemeControllerState['themeVariables']) {
     state.themeVariables = { ...state.themeVariables, ...themeVariables }
+    const themeVariablesSnapshot = ThemeController.getSnapshot().themeVariables
+
     try {
       const emailConnector = ConnectorController.getEmailConnector()
+
       if (emailConnector) {
         emailConnector.provider.syncTheme({
-          themeVariables: ThemeController.getSnapshot().themeVariables,
-          w3mThemeVariables: getCommonThemeVariables(
-            ThemeController.getSnapshot().themeVariables,
-            state.themeMode
-          )
+          themeVariables: themeVariablesSnapshot,
+          w3mThemeVariables: getW3mThemeVariables(themeVariablesSnapshot, state.themeMode)
         })
       }
     } catch {
